requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
numpy
jinja2
python-multipart
